socket architeture changed,create folder,ignore messagf,stored receiveid
Channels is now an abstract class, since run and send were equal to all channels
A peer now always ignores putchunk messages from itself.
Now when a chunk is to be stored the app creates a folder(if it doesnt exit) with the name Peer <ID> to store.
Stored messages are sent after storing a chunk

Still need to check the replicationDegree, all Peers except the iniator stores all chunks requested